[Bittrex] Move BittrexOrder from Account to Trade
